Cast sector size

+++ dvd_cell.c
@@ -51,7 +51,7 @@ ssize_t dvd_cell_blocks(const ifo_handle_t *vmg_ifo, const ifo_handle_t *vts_ifo
 	last_sector = dvd_cell_last_sector(vmg_ifo, vts_ifo, track_number, cell_number);
 
 	ssize_t blocks = 0;
-	blocks = last_sector - first_sector;
+	blocks = (ssize_t)(last_sector - first_sector);
 
 	// Include the last cell
 	if(last_sector == first_sector)
